refactor: unify payload framing under wireFormat (drop serializationForm)
Both schema-source paths now derive the Avro payload offset from a single
WireFormatExtractor, so framing is owned by one concept regardless of how the
schema is chosen:

- Add WireFormat.NONE (bare Avro, offset 0) + NoneWireFormatExtractor.
- EXPRESSION resolver now uses the wire-format extractor for the offset (instead
  of serializationForm), supporting Confluent/Apicurio/header/none envelopes.
- Remove the SerializationForm enum and config field; schema-form replaces it
  with wireFormat (incl. None) and a flat schemaIdSource + EL mappings.
